commands/update: do validation for update with --rev inside Stage
