+++ dvc/repo/update.py
@@ -1,5 +1,3 @@
-from dvc.exceptions import UpdateWithRevNotPossibleError
-
 from . import locked
 
 
@@ -8,9 +6,6 @@ def update(self, target, rev=None):
     from dvc.stage import Stage
 
     stage = Stage.load(self, target)
-
-    if not stage.is_repo_import and rev:
-        raise UpdateWithRevNotPossibleError()
 
     stage.update(rev=rev)
